Re-added nocopy support

+++ src/main/java/org/gestore/plugin/source/fastaSource.java
@@ -5,6 +5,7 @@ import org.gestore.getfasta;
 import org.apache.hadoop.fs.*; 
 import java.util.*; 
 import java.io.File;
+import org.apache.commons.lang.ArrayUtils;
 
 public class fastaSource implements sourceType{
     public fastaSource() {
@@ -35,41 +36,46 @@ public class fastaSource implements sourceType{
         FileStatus[] resultFilesExists = fs.globStatus(new Path(temp_dir + "/existing-r-*"));
         FileStatus[] resultFilesDeleted = fs.globStatus(new Path(temp_dir + "/deleted-r-*"));
         FileStatus[] resultFilesMetadata = fs.globStatus(new Path(temp_dir + "/metadata-r-*"));
+	FileStatus[] allFiles = (FileStatus [])ArrayUtils.addAll(resultFilesExists, resultFilesDeleted);
+	allFiles = (FileStatus [])ArrayUtils.addAll(allFiles, resultFilesExists);
 
+	if(!params.get("copy").equals("true")) {
+		return allFiles;
 
-        int fileCounter = 0;
-
-        for(FileStatus file : resultFilesExists) {
-            if(Integer.parseInt(params.get("split")) == 1) {
-                fs.copyToLocalFile(true, file.getPath(), new Path(real_temp_path));
-            } else {
-                fs.copyToLocalFile(true, file.getPath(), new Path(real_temp_path + "." + Integer.toString(fileCounter)));
-            }
-            fileCounter++;
-        }
-        fileCounter = 0;
-        for(FileStatus file : resultFilesDeleted) {
-            if(Integer.parseInt(params.get("split")) == 1 ) {
-                fs.copyToLocalFile(true, file.getPath(), new Path(real_temp_path + ".deleted"));
-            } else {
-                fs.copyToLocalFile(true, file.getPath(), new Path(real_temp_path + ".deleted." + Integer.toString(fileCounter)));
-            }
-            fileCounter++;
-        }
-        fileCounter = 0;
-        for(FileStatus file : resultFilesMetadata) {
-            if(Integer.parseInt(params.get("split")) == 1 ) {
-                fs.copyToLocalFile(true, file.getPath(), new Path(real_temp_path + ".metadata"));
-            } else {
-                fs.copyToLocalFile(true, file.getPath(), new Path(real_temp_path + ".metadata." + Integer.toString(fileCounter)));
-            }
-            fileCounter++;
-        }
+	} else {
+		local_file_system.delete(new Path(real_temp_path + "*"), true);
+	        int fileCounter = 0;
+	
+        	for(FileStatus file : resultFilesExists) {
+            		if(Integer.parseInt(params.get("split")) == 1) {
+                		fs.copyToLocalFile(true, file.getPath(), new Path(real_temp_path));
+			} else {
+                		fs.copyToLocalFile(true, file.getPath(), new Path(real_temp_path + "." + Integer.toString(fileCounter)));
+            		}
+            		fileCounter++;
+        	}
+        	fileCounter = 0;
+        	for(FileStatus file : resultFilesDeleted) {
+            		if(Integer.parseInt(params.get("split")) == 1 ) {
+                		fs.copyToLocalFile(true, file.getPath(), new Path(real_temp_path + ".deleted"));
+            		} else {
+                		fs.copyToLocalFile(true, file.getPath(), new Path(real_temp_path + ".deleted." + Integer.toString(fileCounter)));
+            		}
+            	fileCounter++;
+        	}
+        	fileCounter = 0;
+        	for(FileStatus file : resultFilesMetadata) {
+            		if(Integer.parseInt(params.get("split")) == 1 ) {
+                		fs.copyToLocalFile(true, file.getPath(), new Path(real_temp_path + ".metadata"));
+            		} else {
+                		fs.copyToLocalFile(true, file.getPath(), new Path(real_temp_path + ".metadata." + Integer.toString(fileCounter)));
+            		}
+            		fileCounter++;
+        	}
         
-        local_file_system.delete(new Path(real_temp_path + "*"), true);
         
-        
-        fs.delete(new Path(temp_dir), true);
+        	fs.delete(new Path(temp_dir), true);
+	}
         
         FileStatus[] files = local_file_system.globStatus(new Path(real_temp_path + "*"));
         return files;
